+++ src/app/common/components/targeting/targeting-custom-option-input/targeting-custom-option-input.component.ts
@@ -3,6 +3,8 @@ import { COMMA, ENTER, SPACE } from '@angular/cdk/keycodes';
 import { MatChipInputEvent } from '@angular/material/chips';
 import { TargetingOption, TargetingOptionType, TargetingOptionValue } from 'models/targeting-option.model';
 import { prepareCustomOption } from 'common/components/targeting/targeting.helpers2';
+import { DecentralandConverter } from 'common/utilities/targeting-converter/decentraland-converter';
+import { CryptovoxelsConverter } from 'common/utilities/targeting-converter/cryptovoxels-converter';
 import { faPlus } from '@fortawesome/free-solid-svg-icons';
 
 @Component({
@@ -48,7 +50,13 @@ export class TargetingCustomOptionInputComponent {
 
   private adjustValueBeforeSave(value: string): TargetingOptionValue {
     const trimmedValue = value.trim().split(' ').join('').toLowerCase();
-    const option = prepareCustomOption(trimmedValue, this.option.id);
+    let url;
+    if (CryptovoxelsConverter.ID === this.option.id) {
+      url = this.prepareCryptovoxelsUrl(trimmedValue);
+    } else if ('site/domain' === this.option.id) {
+      url = `https://${trimmedValue}`;
+    }
+    const option = prepareCustomOption(trimmedValue, this.option.id, url);
     option.selected = true;
     return option;
   }
@@ -67,8 +75,21 @@ export class TargetingCustomOptionInputComponent {
 
   saveCustomParcel(coordinateX: string, coordinateY: string): void {
     const value = `(${coordinateX}, ${coordinateY})`;
-    const option = prepareCustomOption(value, this.option.id);
+    const url = DecentralandConverter.ID === this.option.id ? this.prepareDecentralandUrl(value) : undefined;
+    const option = prepareCustomOption(value, this.option.id, url);
     this.customOptionsArray.push(option);
     this.itemsChange.emit(this.customOptionsArray);
+  }
+
+  prepareCryptovoxelsUrl(value: string): string {
+    const converter = new CryptovoxelsConverter();
+    const backendUrl = `https://${converter.encodeValue(value)}`;
+    return converter.convertBackendUrlToValidUrl(backendUrl);
+  }
+
+  prepareDecentralandUrl(value: string): string {
+    const converter = new DecentralandConverter();
+    const backendUrl = `https://${converter.encodeValue(value)}`;
+    return converter.convertBackendUrlToValidUrl(backendUrl);
   }
 }