+++ src/app/common/components/targeting/targeting-display/targeting-display.component.ts
@@ -21,6 +21,7 @@ export class TargetingDisplayComponent implements OnChanges {
     chosenTargeting: {
       id: string;
       label: string;
+      url: string | undefined;
     }[];
   }[];
   faClose = faClose;
@@ -36,6 +37,7 @@ export class TargetingDisplayComponent implements OnChanges {
       const chosenTargetingItem = {
         id: item.id,
         label: label,
+        url: item.url,
       };
 
       const viewModelParentPathIndex = this.viewModel.findIndex(viewModelItem => viewModelItem.parentPath === path);