+++ CHANGELOG.md
@@ -8,12 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Site's statuses (pending approval, rejected)
+- Targeting chip representing domain routes to page on click
 ### Changed
 - Load direct link ad content during campaign load
 ### Fixed
 - Set chart's date filter
 - Sites' order on dashboard after status change
 - Set conversion's mutable value
+- Handle HTTP 403 error
 
 ## [2.2.4] - 2022-12-29
 ### Fixed